Place prosemirror commands in separate file.
