+++ package.json
@@ -1,89 +1,88 @@
 {
-    "name": "vid-annot-slate",
-    "version": "0.1.0",
-    "private": true,
-    "dependencies": {
-        "@emotion/core": "10.0.10",
-        "@emotion/styled": "10.0.10",
-        "@kenshooui/react-multi-select": "^1.1.3",
-        "@material-ui/core": "4.1.1",
-        "@material-ui/icons": "4.2.0",
-        "@material-ui/lab": "^4.0.0-alpha.16",
-        "@slate-editor/font-size-plugin": "^5.1.0",
-        "@types/classnames": "2.2.8",
-        "@types/jest": "24.0.15",
-        "@types/node": "12.0.8",
-        "@types/prosemirror-commands": "1.0.1",
-        "@types/prosemirror-inputrules": "1.0.2",
-        "@types/prosemirror-keymap": "1.0.1",
-        "@types/prosemirror-state": "1.2.3",
-        "@types/react": "16.8.21",
-        "@types/react-dom": "16.8.4",
-        "@types/react-router-dom": "4.3.4",
-        "@types/seedrandom": "2.4.28",
-        "@typescript-eslint/eslint-plugin": "1.10.2",
-        "@typescript-eslint/parser": "1.10.2",
-        "babel-eslint": "^10.0.1",
-        "classnames": "2.2.6",
-        "create-react-class": "^15.6.3",
-        "dropbox": "^4.0.17",
-        "eslint": "^5.16.0",
-        "eslint-config-prettier": "^4.3.0",
-        "eslint-config-react-app": "^4.0.1",
-        "eslint-plugin-flowtype": "2.50.3",
-        "eslint-plugin-import": "^2.17.3",
-        "eslint-plugin-jsx-a11y": "^6.2.1",
-        "eslint-plugin-react": "^7.13.0",
-        "eslint-plugin-react-hooks": "^1.6.0",
-        "fromentries": "^1.1.0",
-        "get-youtube-id": "^1.0.1",
-        "get-youtube-title": "^1.0.0",
-        "history": "4.9.0",
-        "immutable": "^4.0.0-rc.12",
-        "is-hotkey": "0.1.4",
-        "lodash": "^4.17.11",
-        "material-ui-popup-state": "1.3.2",
-        "popper": "^1.0.1",
-        "prismjs": "1.16.0",
-        "prosemirror-example-setup": "^1.0.1",
-        "prosemirror-inputrules": "^1.0.4",
-        "prosemirror-keymap": "^1.0.1",
-        "prosemirror-markdown": "^1.3.0",
-        "prosemirror-model": "^1.7.1",
-        "prosemirror-schema-basic": "^1.0.1",
-        "prosemirror-tables": "^0.8.1",
-        "prosemirror-utils": "^0.9.2",
-        "prosemirror-view": "^1.9.8",
-        "querystring": "^0.2.0",
-        "react": "16.8.6",
-        "react-contexify": "4.1.0",
-        "react-dom": "16.8.6",
-        "react-jss": "^8.6.1",
-        "react-modal": "^3.8.1",
-        "react-router": "5.0.0",
-        "react-router-dom": "5.0.0",
-        "react-scripts": "3.0.0",
-        "reactjs-popup": "^1.3.2",
-        "redux": "^4.0.1",
-        "slug": "1.1.0",
-        "typescript": "^3.4.5"
-    },
-    "scripts": {
-        "eslint": "eslint \"src/**/*.js{,x}\" && eslint \"extension/*.js\"",
-        "start": "yarn eslint && react-scripts start",
-        "build": "yarn eslint && react-scripts build",
-        "test": "yarn eslint && react-scripts test",
-        "eject": "react-scripts eject",
-        "build_ext": "cd extension; yarn build; cd ..",
-        "build_ext_and_start": "concurrently \"yarn build_ext\" \"yarn start\""
-    },
-    "eslintConfig": {
-        "extends": "react-app"
-    },
-    "browserslist": [
-        ">0.2%",
-        "not dead",
-        "not ie <= 11",
-        "not op_mini all"
-    ]
+	"name": "vid-annot-slate",
+	"version": "0.1.0",
+	"private": true,
+	"dependencies": {
+		"@emotion/core": "10.0.10",
+		"@emotion/styled": "10.0.10",
+		"@kenshooui/react-multi-select": "^1.1.3",
+		"@material-ui/core": "4.1.1",
+		"@material-ui/icons": "4.2.0",
+		"@material-ui/lab": "^4.0.0-alpha.16",
+		"@slate-editor/font-size-plugin": "^5.1.0",
+		"@types/classnames": "2.2.8",
+		"@types/jest": "24.0.15",
+		"@types/node": "12.0.8",
+		"@types/prosemirror-commands": "1.0.1",
+		"@types/prosemirror-inputrules": "1.0.2",
+		"@types/prosemirror-keymap": "1.0.1",
+		"@types/prosemirror-state": "1.2.3",
+		"@types/react": "16.8.21",
+		"@types/react-dom": "16.8.4",
+		"@types/react-router-dom": "4.3.4",
+		"@types/seedrandom": "2.4.28",
+		"@typescript-eslint/eslint-plugin": "1.10.2",
+		"@typescript-eslint/parser": "1.10.2",
+		"babel-eslint": "^10.0.1",
+		"classnames": "2.2.6",
+		"create-react-class": "^15.6.3",
+		"dropbox": "^4.0.17",
+		"eslint": "^5.16.0",
+		"eslint-config-prettier": "^4.3.0",
+		"eslint-config-react-app": "^4.0.1",
+		"eslint-plugin-flowtype": "2.50.3",
+		"eslint-plugin-import": "^2.17.3",
+		"eslint-plugin-jsx-a11y": "^6.2.1",
+		"eslint-plugin-react": "^7.13.0",
+		"eslint-plugin-react-hooks": "^1.6.0",
+		"get-youtube-id": "^1.0.1",
+		"get-youtube-title": "^1.0.0",
+		"history": "4.9.0",
+		"immutable": "^4.0.0-rc.12",
+		"is-hotkey": "0.1.4",
+		"lodash": "^4.17.11",
+		"material-ui-popup-state": "1.3.2",
+		"popper": "^1.0.1",
+		"prismjs": "1.16.0",
+		"prosemirror-example-setup": "^1.0.1",
+		"prosemirror-inputrules": "^1.0.4",
+		"prosemirror-keymap": "^1.0.1",
+		"prosemirror-markdown": "^1.3.0",
+		"prosemirror-model": "^1.7.1",
+		"prosemirror-schema-basic": "^1.0.1",
+		"prosemirror-tables": "^0.8.1",
+		"prosemirror-utils": "^0.9.2",
+		"prosemirror-view": "^1.9.8",
+		"querystring": "^0.2.0",
+		"react": "16.8.6",
+		"react-contexify": "4.1.0",
+		"react-dom": "16.8.6",
+		"react-jss": "^8.6.1",
+		"react-modal": "^3.8.1",
+		"react-router": "5.0.0",
+		"react-router-dom": "5.0.0",
+		"react-scripts": "3.0.0",
+		"reactjs-popup": "^1.3.2",
+		"redux": "^4.0.1",
+		"slug": "1.1.0",
+		"typescript": "^3.5.2"
+	},
+	"scripts": {
+		"eslint": "eslint \"src/**/*.js{,x}\" && eslint \"extension/*.js\"",
+		"start": "yarn eslint && react-scripts start",
+		"build": "yarn eslint && react-scripts build",
+		"test": "yarn eslint && react-scripts test",
+		"eject": "react-scripts eject",
+		"build_ext": "cd extension; yarn build; cd ..",
+		"build_ext_and_start": "concurrently \"yarn build_ext\" \"yarn start\""
+	},
+	"eslintConfig": {
+		"extends": "react-app"
+	},
+	"browserslist": [
+		">0.2%",
+		"not dead",
+		"not ie <= 11",
+		"not op_mini all"
+	]
 }
